+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 # Changelog
 ## __WORK IN PROGRESS__
 * Add refactor to replace options with min/max value
+* Fix: Enable/Disable template refactor preserves `$if` conditions
 
 ## 0.0.14 (2023-10-09)
 * Suggest refactors when a matching enable/disable template is found